+++ src/main/java/com/nguyengiap/security/model/request_model/BankingRequestOTP.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class BankingRequestOTP {
     private String toAccount;
-    private Integer fund;
+    private double fund;
     private String message;
     private String otp;
 }
